Personalizing zsh configs and Documentation
I adapted the configurations to my personal liking.

I documented the `zsh` configuration. I guess I'll know if I did a
good job if in a couple of months this guidelines help me save
a lot of time reviewing the references. Nevertheless, I also included
the references and relevant links.